fixup: Polished Progress counter types
Most current use cases might be borderline OK with "int" counters
because of sfileno limits, but those limits may be raised in the future,
db _slot_ counters are technically not subject to sfileno limits,
storeRebuildProgress() already uses doubles to sum things up (probably
because int64_t was not easily accessible at the time), and known future
users (e.g., store_rebuild_progress) will probably need more bits.
Reduce anticipated future changes while adding this new class.

+++ src/store_rebuild.h
@@ -40,14 +40,14 @@ public:
 class Progress
 {
 public:
-    Progress(const int stepsCompleted, const int stepsTotal):
+    Progress(const int64_t stepsCompleted, const int64_t stepsTotal):
         completed(stepsCompleted), goal(stepsTotal) {}
 
     /// brief progress report suitable for level-0/1 debugging
     void print(std::ostream &os) const;
 
-    int completed; ///< the number of finished work steps
-    int goal; ///< the known total number of work steps (or negative)
+    int64_t completed; ///< the number of finished work steps
+    int64_t goal; ///< the known total number of work steps (or negative)
 };
 
 inline
